Add ProjectService and update ProjectPageDTO
Introduced ProjectService with dependency injection for ProjectRepository. Changed totalRecords type in ProjectPageDTO from Long to Integer for consistency with other fields.

+++ src/main/java/com/solarize/solarizeWebBackend/modules/project/dto/ProjectPageDTO.java
@@ -12,7 +12,7 @@ import java.util.List;
 @AllArgsConstructor
 public class ProjectPageDTO {
 
-    private Long totalRecords;
+    private Integer totalRecords;
     private Integer totalPages;
     private Integer totalResults;
     private List<ProjectSummaryDTO> projects;
